Improved name and description
This is how my local version is done (except for the urls and web_url). In a designer's eye, "Tumblr" looks prettier (and it is good grammar) than "tumblr." (Tumblr itself throughout their documentation use "Tumblr" instead of "tumblr.").

The second change is the description. Again, grammar.  (A period at the end... it looks prettier this way in the webstore).